sched: Change task_struct::state
Change the type and name of task_struct::state. Drop the volatile and
shrink it to an 'unsigned int'. Rename it in order to find all uses
such that we can use READ_ONCE/WRITE_ONCE as appropriate.

Acked-by: Will Deacon <will@kernel.org>
Acked-by: Daniel Thompson <daniel.thompson@linaro.org>
Link: https://lore.kernel.org/r/20210611082838.550736351@infradead.org

+++ block/blk-mq.c
@@ -3222,7 +3222,7 @@ static bool blk_mq_poll_hybrid_sleep(struct request_queue *q,
 static bool __blk_mq_poll(struct blk_mq_hw_ctx *hctx, struct request *rq)
 {
 	struct request_queue *q = hctx->queue;
-	long state;
+	unsigned int state;
 
 	/*
 	 * If we sleep, have the caller restart the poll loop to reset
